Initial release
Fixed the not working output -> Default focal length for GoPro was wrong

Updated examples and readme

Added a grid overlay:
When grid=True then a grid of blue lines is put over the input source before
applying the lenscorrection.
This helps to see, what acually is done to the image.